Moved FormsStreamService to common folder
