+++ frontend/src/employee/views/FormView.jsx
@@ -9,7 +9,7 @@ import { FaTrash } from 'react-icons/fa';
 
 import authService from '../services/AuthService';
 import formService from '../services/FormService';
-import formStreamService from '../services/FormsStreamService';
+import formStreamService from '../../common/services/FormsStreamService';
 import deviceStreamService from '../services/DeviceStreamService';
 import dataValidator from '../../common/helpers/DataValidator';
