+++ frontend/src/employee/views/FormsListView.jsx
@@ -12,7 +12,7 @@ import employeeService from '../services/EmployeeService';
 import formService from '../services/FormService';
 import schemaService from '../services/SchemaService';
 import deviceStreamService from '../services/DeviceStreamService';
-import formStreamService from '../services/FormsStreamService';
+import formStreamService from '../../common/services/FormsStreamService';
 
 import LoadingView from '../../common/views/utility/LoadingView';
