+++ frontend/src/patient/views/EndView.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { Container } from 'react-bootstrap';
 import { useHistory } from 'react-router-dom';
-import formStreamService from '../services/FormsStreamService';
+import formStreamService from '../../common/services/FormsStreamService';
 
 const EndView = () => {
   const history = useHistory();
